We are able to reproduce the issue https://github.com/mpeg5/xeve/issues/108 with the following build command: ``` cmake -DCMAKE_C_FLAGS="-mavx" .. make ``` Encode command: ``` ./xeve_app -i <input.y4m> -o <output.evc> ```
Error message:
```
XEVE: eXtra-fast Essential Video Encoder
Segmentation fault
```

On passing "-mavx2" as C_Flag (cmake -DCMAKE_C_FLAGS="-mavx" ..), it sets the flag
for all files in the project, including the source C files. So, the compiler is
creating optimizations which are not handled properly resulting in a segmentation fault.

This PR fixes the issue by updating the CmakeLists.txt files.